tui: extract engine-free tui-dashboard-core so the aggregator drops libghostty
The tui-dashboard aggregator linked libghostty-vt.so (via tui -> ix-vt) even
though it only renders frames and never drives a PTY. Feature-gating ix-vt does
not help: the workspace builds with cargo --workspace, which unifies tui's
features, so tui-py/tui-node force the engine on the shared tui build and the
aggregator links it anyway.

Extract the engine-free surface the aggregator uses into a new tui-dashboard-core
crate (the wire types TerminalFrame/ProducerSnapshot/socket_dir/socket_path and
the browser server Hub/serve_hub/Dashboard). tui-dashboard now depends on that
crate instead of tui, so it no longer pulls the engine. tui keeps the engine and
re-exports the moved types, so its public API and tui-py/tui-node/mcp are
unchanged. collect_frames and the in-process serve stay in tui.

Verified: tui-dashboard's binary no longer lists libghostty-vt in NEEDED;
rust-package-tests and lint pass.
