+++ packages/tui-dashboard-core/package.nix
@@ -1,0 +1,4 @@
+{
+  id = "tui-dashboard-core";
+  inRustWorkspace = true;
+}
