+++ packages/tui-dashboard-core/src/dashboard/dashboard.html
