+++ packages/tui-dashboard/Cargo.toml
@@ -16,4 +16,4 @@ path = "src/main.rs"
 clap = { workspace = true, features = ["derive"] }
 serde_json.workspace = true
 tokio = { workspace = true, features = ["macros", "rt-multi-thread", "net", "io-util", "time", "signal"] }
-tui = { workspace = true, features = ["dashboard"] }
+tui-dashboard-core.workspace = true
