+++ packages/tui/src/lib.rs
@@ -48,10 +48,12 @@ mod slice;
 mod types;
 
 #[cfg(feature = "dashboard")]
-pub use dashboard::{Dashboard, Hub, serve, serve_hub};
+pub use dashboard::serve;
 pub use error::{Error, Result};
 #[cfg(any(feature = "dashboard", feature = "publish"))]
-pub use frame::{ProducerSnapshot, TerminalFrame, socket_dir, socket_path};
+pub use tui_dashboard_core::{ProducerSnapshot, TerminalFrame, socket_dir, socket_path};
+#[cfg(feature = "dashboard")]
+pub use tui_dashboard_core::{Dashboard, Hub, serve_hub};
 pub use manager::{TuiInstance, TuiManager};
 #[cfg(feature = "publish")]
 pub use publish::{Publisher, publish};
